+++ app/src/main/java/com/android/academy/fundamentals/workshop_3/WS03CoroutinesProblemFragment.kt
@@ -1,4 +1,6 @@
-package com.android.academy.fundamentals.workshop03.solution
+@file:Suppress("unused")
+
+package com.android.academy.fundamentals.workshop_3
 
 import android.os.Bundle
 import android.view.View
@@ -8,7 +10,7 @@ import androidx.fragment.app.Fragment
 import com.android.academy.fundamentals.R
 import kotlinx.coroutines.*
 
-class WS03CoroutinesSolutionFragment : Fragment(R.layout.fragment_coroutines_scope_cancel) {
+class WS03CoroutinesProblemFragment : Fragment(R.layout.fragment_coroutines_scope_cancel) {
     // This is exception handler that will print caught errors to log
     private val exceptionHandler = CoroutineExceptionHandler { coroutineContext, exception ->
         println("CoroutineExceptionHandler got $exception in $coroutineContext")
@@ -45,23 +47,13 @@ class WS03CoroutinesSolutionFragment : Fragment(R.layout.fragment_coroutines_sco
     private fun startCoroutines() {
         toggleButtons(true)
 
-        // Run odd number producing coroutine
-        scope.launch {
-            runOddsCoroutine()
-        }
-        // Run negatives number producing coroutine
-        scope.launch {
-            runNegativesCoroutine()
-        }
-        // Run mod by two result from number coroutine
-        // Run it with GlobalScope to observe cancellation ignorance
-        GlobalScope.launch {
-            runModByTwoCoroutine()
-        }
-        // Run coroutine that fails after a second
-        scope.launch {
-            runCoroutineThatFails()
-        }
+        // Add run odd number producing coroutine via launch in scope
+
+        // Add run negatives number producing coroutine via launch in scope
+
+        // Add run mod by two result from number coroutine via launch in GlobalScope
+
+        // Add run coroutine that fails after a second
     }
 
     // Coroutine that produces odd numbers every second
@@ -102,15 +94,15 @@ class WS03CoroutinesSolutionFragment : Fragment(R.layout.fragment_coroutines_sco
         throw IllegalStateException("Some exception")
     }
 
-    private suspend fun showResult(text: String, resultView: TextView?) {
-        withContext(Dispatchers.Main) {
-            resultView?.text = text
-        }
+    private fun showResult(text: String, resultView: TextView?) {
+        // Fix access of UI component from main dispatcher
+        // Don't forget about suspend keyword
+        resultView?.text = text
     }
 
     private fun cancelCoroutines() {
-        // Cancel all current jobs via parent job
-        scope.cancel()
+        // Add cancellation of all current jobs via parent job
+
         // Set new scope with fresh SupervisorJob after cancel
         scope = CoroutineScope(SupervisorJob() + Dispatchers.Default + exceptionHandler)
